refactor : 임시 필드 영속성 대상 제거
