+++ src/main/java/usw/suwiki/domain/lecture/controller/dto/LectureDetailResponseDto.java
@@ -27,7 +27,7 @@ public class LectureDetailResponseDto {
         this.id = entity.getId();
         this.semesterList = entity.getSemester();
         this.professor = entity.getProfessor();
-        this.lectureType = entity.getLectureDetail().getType();
+        this.lectureType = entity.getType();
         this.lectureName = entity.getName();
         this.majorType = entity.getMajorType();
         this.lectureTotalAvg = entity.getLectureAverage().getLectureTotalAvg();
