+++ src/main/java/usw/suwiki/domain/lecture/domain/repository/LectureRepository.java
@@ -1,4 +1,4 @@
-package usw.suwiki.domain.lecture.repository;
+package usw.suwiki.domain.lecture.domain.repository;
 
 import usw.suwiki.domain.lecture.domain.Lecture;
 import usw.suwiki.domain.lecture.controller.dto.LectureFindOption;
